--- a/edwards25519/field_amd64.s
+++ b/edwards25519/field_amd64.s
@@ -1,4 +1,4 @@
-//+build amd64
+//+build amd64 AND !go1.12
 
 #include "textflag.h"