--- a/edwards25519/field_mul64.go
+++ b/edwards25519/field_mul64.go
@@ -54,6 +54,10 @@ func (fe *FieldElement) Mul(a, b *FieldElement) *FieldElement {
 	c1l, carry = bits.Add64(c1l, l, 0)
 	c1h, _ = bits.Add64(c1h, h, carry)
 
+	c1l, carry = bits.Add64((c0l>>51)|(c0h<<13), c1l, 0)
+	c1h, _ = bits.Add64(c1h, 0, carry)
+	h0 := c0l & 0x7ffffffffffff
+
 	c2h, c2l := bits.Mul64(b0, a2)
 	h, l = bits.Mul64(b1, a1)
 	c2l, carry = bits.Add64(c2l, l, 0)
@@ -67,6 +71,10 @@ func (fe *FieldElement) Mul(a, b *FieldElement) *FieldElement {
 	h, l = bits.Mul64(b4_19, a3)
 	c2l, carry = bits.Add64(c2l, l, 0)
 	c2h, _ = bits.Add64(c2h, h, carry)
+
+	c2l, carry = bits.Add64((c1l>>51)|(c1h<<13), c2l, 0)
+	c2h, _ = bits.Add64(c2h, 0, carry)
+	h1 := c1l & 0x7ffffffffffff
 
 	c3h, c3l := bits.Mul64(b0, a3)
 	h, l = bits.Mul64(b1, a2)
@@ -82,6 +90,10 @@ func (fe *FieldElement) Mul(a, b *FieldElement) *FieldElement {
 	c3l, carry = bits.Add64(c3l, l, 0)
 	c3h, _ = bits.Add64(c3h, h, carry)
 
+	c3l, carry = bits.Add64((c2l>>51)|(c2h<<13), c3l, 0)
+	c3h, _ = bits.Add64(c3h, 0, carry)
+	h2 := c2l & 0x7ffffffffffff
+
 	c4h, c4l := bits.Mul64(b0, a4)
 	h, l = bits.Mul64(b1, a3)
 	c4l, carry = bits.Add64(c4l, l, 0)
@@ -95,18 +107,6 @@ func (fe *FieldElement) Mul(a, b *FieldElement) *FieldElement {
 	h, l = bits.Mul64(b4, a0)
 	c4l, carry = bits.Add64(c4l, l, 0)
 	c4h, _ = bits.Add64(c4h, h, carry)
-
-	c1l, carry = bits.Add64((c0l>>51)|(c0h<<13), c1l, 0)
-	c1h, _ = bits.Add64(c1h, 0, carry)
-	h0 := c0l & 0x7ffffffffffff
-
-	c2l, carry = bits.Add64((c1l>>51)|(c1h<<13), c2l, 0)
-	c2h, _ = bits.Add64(c2h, 0, carry)
-	h1 := c1l & 0x7ffffffffffff
-
-	c3l, carry = bits.Add64((c2l>>51)|(c2h<<13), c3l, 0)
-	c3h, _ = bits.Add64(c3h, 0, carry)
-	h2 := c2l & 0x7ffffffffffff
 
 	c4l, carry = bits.Add64((c3l>>51)|(c3h<<13), c4l, 0)
 	c4h, _ = bits.Add64(c4h, 0, carry)
@@ -155,6 +155,10 @@ func (fe *FieldElement) Square(a *FieldElement) *FieldElement {
 	c1l, carry = bits.Add64(c1l, l, 0)
 	c1h, _ = bits.Add64(c1h, h, carry)
 
+	c1l, carry = bits.Add64((c0l>>51)|(c0h<<13), c1l, 0)
+	c1h, _ = bits.Add64(c1h, 0, carry)
+	h0 := c0l & 0x7ffffffffffff
+
 	c2h, c2l := bits.Mul64(a0, 2*a2)
 	h, l = bits.Mul64(a1, a1)
 	c2l, carry = bits.Add64(c2l, l, 0)
@@ -162,6 +166,10 @@ func (fe *FieldElement) Square(a *FieldElement) *FieldElement {
 	h, l = bits.Mul64(a4_19, a3*2)
 	c2l, carry = bits.Add64(c2l, l, 0)
 	c2h, _ = bits.Add64(c2h, h, carry)
+
+	c2l, carry = bits.Add64((c1l>>51)|(c1h<<13), c2l, 0)
+	c2h, _ = bits.Add64(c2h, 0, carry)
+	h1 := c1l & 0x7ffffffffffff
 
 	c3h, c3l := bits.Mul64(a0, 2*a3)
 	h, l = bits.Mul64(2*a1, a2)
@@ -171,6 +179,10 @@ func (fe *FieldElement) Square(a *FieldElement) *FieldElement {
 	c3l, carry = bits.Add64(c3l, l, 0)
 	c3h, _ = bits.Add64(c3h, h, carry)
 
+	c3l, carry = bits.Add64((c2l>>51)|(c2h<<13), c3l, 0)
+	c3h, _ = bits.Add64(c3h, 0, carry)
+	h2 := c2l & 0x7ffffffffffff
+
 	c4h, c4l := bits.Mul64(a0, 2*a4)
 	h, l = bits.Mul64(a2, a2)
 	c4l, carry = bits.Add64(c4l, l, 0)
@@ -178,18 +190,6 @@ func (fe *FieldElement) Square(a *FieldElement) *FieldElement {
 	h, l = bits.Mul64(a3, 2*a1)
 	c4l, carry = bits.Add64(c4l, l, 0)
 	c4h, _ = bits.Add64(c4h, h, carry)
-
-	c1l, carry = bits.Add64((c0l>>51)|(c0h<<13), c1l, 0)
-	c1h, _ = bits.Add64(c1h, 0, carry)
-	h0 := c0l & 0x7ffffffffffff
-
-	c2l, carry = bits.Add64((c1l>>51)|(c1h<<13), c2l, 0)
-	c2h, _ = bits.Add64(c2h, 0, carry)
-	h1 := c1l & 0x7ffffffffffff
-
-	c3l, carry = bits.Add64((c2l>>51)|(c2h<<13), c3l, 0)
-	c3h, _ = bits.Add64(c3h, 0, carry)
-	h2 := c2l & 0x7ffffffffffff
 
 	c4l, carry = bits.Add64((c3l>>51)|(c3h<<13), c4l, 0)
 	c4h, _ = bits.Add64(c4h, 0, carry)