It's ~5% slower, but portable!

Old:

```
BenchmarkVarTimeScalarMult-8                  	   30000	     48631 ns/op
BenchmarkScalarMult-8                         	   20000	     75398 ns/op
BenchmarkScalarMultTableCompute-8             	    2000	   1075480 ns/op
BenchmarkScalarMultTableScalarMult-8          	  100000	     23533 ns/op
BenchmarkScalarMultTableVarTimeScalarMult-8   	  100000	     11542 ns/op
```

New:
```
BenchmarkVarTimeScalarMult-8                  	   30000	     51993 ns/op
BenchmarkScalarMult-8                         	   20000	     79343 ns/op
BenchmarkScalarMultTableCompute-8             	    2000	   1067662 ns/op
BenchmarkScalarMultTableScalarMult-8          	   50000	     25068 ns/op
BenchmarkScalarMultTableVarTimeScalarMult-8   	  100000	     12851 ns/op
```

[Generated assembly](https://godbolt.org/z/sQSPqf)