@bwesterb 2019-01-14
Currently still prohibitively slower

Old:

```
BenchmarkVarTimeScalarMult-8                  	   30000	     48631 ns/op
BenchmarkScalarMult-8                         	   20000	     75398 ns/op
BenchmarkScalarMultTableCompute-8             	    2000	   1075480 ns/op
BenchmarkScalarMultTableScalarMult-8          	  100000	     23533 ns/op
BenchmarkScalarMultTableVarTimeScalarMult-8   	  100000	     11542 ns/op
```

New:
```
BenchmarkVarTimeScalarMult-8                  	   20000	     77729 ns/op
BenchmarkScalarMult-8                         	   10000	    108805 ns/op
BenchmarkScalarMultTableCompute-8             	    1000	   1607489 ns/op
BenchmarkScalarMultTableScalarMult-8          	   50000	     32325 ns/op
BenchmarkScalarMultTableVarTimeScalarMult-8   	  100000	     21184 ns/op
```